Scrapper/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
playwright>=1.40.0
crewai>=0.11.0
supabase>=2.3.0
pydantic>=2.5.2
python-dotenv>=1.0.0
PyYAML>=6.0.1
loguru>=0.7.2
asyncio>=3.4.3 